CCCT-2294 Parse Relearn Tasks For UI
Extracted the parse function to the new data class.
